[Fix #1517] Review comments.

+++ impl/core/src/main/java/io/serverlessworkflow/impl/TaskContextData.java
@@ -40,5 +40,5 @@ public interface TaskContextData {
 
   int iteration();
 
-  short retryAttempt();
+  int retryAttempt();
 }
